Fixed issue with size_t not fitting error codes for fasta_read
Maintained size_t for state and functions that interact with state for simplicity and consistency with sequence limits

+++ src/argparse.h
@@ -33,7 +33,7 @@ typedef struct
 {
     const char *name;
     const char *exts;
-    size_t (*reader)(FILE *, SeqRecord **);
+    int (*reader)(FILE *, SeqRecord **);
 } FormatOption;
 
 typedef struct
